+++ packages/orchestration/index.js
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/export
 export * from './src/types.js';
 export * from './src/utils/address.js';
-export * from './src/utils/tx.js';
+export * from './src/utils/packet.js';
 export * from './src/service.js';
